fixed test cases, corrected db attribute format

+++ backend/app/__init__.py
@@ -41,7 +41,7 @@ def create_app(config=None):
 
     with app.app_context():
         # add the string in run.py to initialize/reset the database
-        #if config == "intialize_db":
+        if config == "intialize_db":
             db.drop_all()
             db.create_all()
             create_locations()
